Simplify definition of "readiness state"
The definition of "maximum active sessions" states:

> [...] must be exactly one for a remote end that is an endpoint node.

This restriction is further supported by the definition of "active
sessions," which states:

> A remote end that is not an intermediary node has at most one active
> session at a given time.

Therefore, the "readiness state" for endpoint nodes can be described in
terms of a precise number ("one") of active sessions. The additional
state concerning the current user prompt is subsumed by this condition.

Simplify the definition of "readiness state" to reflect the limitations
of endpoint nodes.