Go: Correct URL validation
Previously, the condition was described in terms of the union of two
compliments; all string values satisfied this condition. It also
referenced both "absolute URL," and "absolute URL with fragment,"
although the former is a strict subset of the latter. It also misapplied
the term "local scheme" (which describes the "scheme" segment of a URL)
to an expected URL value.

Modify the language to instead describe the intersection of two
compliments. Remove superfluous reference to "absolute URL." Qualify
local URLs using the definition that describes URLs.